Adjustments

+++ daily-exercises/2025-11-6/rep6.c
@@ -3,15 +3,15 @@
 int main(void)
 {
 
-    int number, fact = 1, count = 1;
+    unsigned long int number, fact = 1, count = 1;
     
     printf("Enter a number: ");
-    scanf("%d", &number);
+    scanf("%lu", &number);
     
     while (number < 0)
     {
         printf("Please enter a positive number: ");
-        scanf("%d", &number);
+        scanf("%lu", &number);
     }
 
     while (count <= number)
@@ -21,7 +21,7 @@ int main(void)
         count++;
     }
 
-    printf("The factorial of the number %d is %d", number, fact);
+    printf("The factorial of the number %lu is %lu", number, fact);
 
     return 0;
 }
